Refactor MarkdownRenderer component
Updated the MarkdownRenderer component for improved readability and maintainability. Consider further refactoring as the file is quite lengthy.
[skip gpt_engineer]
